PROPERLY implement super/sub classes.
Also: there are sets of commands that MUST be defined in Report that are later overridden and actually configured in subclasses.  All such **required** methods need to be defined in the superclass, even if they ultimately `raise NotImplementedError`, in order to avoid Unresolved Local problems that may crop up if you don't properly configure the subclasses' overrides of the superclass' methods.